chat_server/requirements.txt
protobuf
grpcio
python-etcd
passlib
bcrypt